rohit commit

+++ day12springjpa/springrestdemo/springrest/src/main/java/com/greatlearning/springrest/dao/EmployeeDAO.java
@@ -8,5 +8,5 @@ import com.greatlearning.springrest.entity.BoxItem;
 public interface EmployeeDAO {
 	public void addText(int fgid, String name, int is_required, int max_length, int min_length);
 
-	void addCheckBoxList(int fgid, String name, int is_required, List<BoxItem> boxlist);
+	void addCheckBoxList(long fgid, String name, int is_required, List<BoxItem> boxlist);
 }
